Fixed issues with GaggleData.getData. There are three cases. First, when we parse data from the gaggled web page, we convert javascript object to Java object(i.e., Namelist, DataMatrix, etc), and stores the result in gaggleData._data.
Another case is the data included in the workflowAction, which is wrapped in FG_GaggleWorkflowDataFromGoose.

Finally, when we reload state, the state data is wrapped in FG_GaggleDeserializedData.

We need to make sure all these scenarios the ._data field point to the same type of objects. So the web handlers can handle them properly.